# docs/swagger.yaml
basePath: /
definitions:
...
  types.APIError:
    properties:
      code:
        type: string
      details:
        type: string
      message:
        type: string
    type: object
  types.APIResponse:
    properties:
      data: {}
      error:
        $ref: '#/definitions/types.APIError'
      success:
        type: boolean
    type: object
  types.AssetInfo:
    properties:
      balance:
        type: string
      balance_wei:
        type: string
      chain_display_name:
        type: string
      chain_id:
        type: integer
      chain_logo_url:
        type: string
      chain_name:
        type: string
      contract_address:
        type: string
      is_native:
        type: boolean
      is_testnet:
        type: boolean
      last_updated:
        type: string
      price_change_24h:
        description: 24小时价格涨跌幅（%）
        type: number
      token_decimals:
        type: integer
      token_logo_url:
        type: string
      token_name:
        type: string
      token_price:
        type: number
      token_symbol:
        type: string
      usd_value:
        type: number
    type: object
...
  types.RefreshTokenRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
...
  types.User:
    properties:
      created_at:
        type: string
      id:
        type: integer
      last_login:
        type: string
      status:
        type: integer
      updated_at:
        type: string
      wallet_address:
        description: 钱包地址作为唯一标识
        type: string
    type: object
  types.UserAssetResponse:
    properties:
      assets:
        description: 所有支持链上的资产，按价值从高到低排序
        items:
          $ref: '#/definitions/types.AssetInfo'
        type: array
      last_updated:
        type: string
      total_usd_value:
        type: number
      wallet_address:
        type: string
    type: object
...
  types.UserProfile:
    properties:
      created_at:
        type: string
      last_login:
        type: string
      wallet_address:
        type: string
    type: object
...
  types.WalletConnectRequest:
    properties:
      message:
        type: string
      signature:
        type: string
      wallet_address:
        type: string
    required:
    - message
    - signature
    - wallet_address
    type: object
  types.WalletConnectResponse:
    properties:
      access_token:
        type: string
      expires_at:
        type: string
      refresh_token:
        type: string
      user:
        $ref: '#/definitions/types.User'
    type: object
host: localhost:8080
info:
  contact: {}
  description: TimeLocker Backend API
  title: TimeLocker Backend API
  version: "1.0"
paths:
...
  /api/v1/assets:
    post:
      consumes:
      - application/json
      description: 获取当前认证用户在所有支持的区块链上的资产信息，包括代币余额、USD价值、价格变化等。如果数据库中没有缓存数据，系统会自动从区块链上刷新最新的资产信息。资产按USD价值从高到低排序。
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取用户资产信息
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/types.UserAssetResponse'
              type: object
        "401":
          description: 未认证或令牌无效
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "500":
          description: 获取资产信息失败或区块链查询错误
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: 获取用户资产列表
      tags:
      - Assets
  /api/v1/assets/refresh:
    post:
      consumes:
      - application/json
      description: 强制从区块链上重新获取用户在所有支持链上的最新资产信息，更新数据库缓存。此操作可能需要较长时间，因为需要查询多个区块链网络的数据。
      produces:
      - application/json
      responses:
        "200":
          description: 资产刷新成功
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                data:
                  type: string
              type: object
        "401":
          description: 未认证或令牌无效
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "500":
          description: 刷新资产失败或区块链查询错误
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: 强制刷新用户资产信息
      tags:
      - Assets
  /api/v1/auth/profile:
    post:
      consumes:
      - application/json
      description: 获取当前认证用户的详细资料信息，包括钱包地址、创建时间等。需要有效的JWT令牌。
      produces:
      - application/json
      responses:
        "200":
          description: 成功获取用户资料
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/types.UserProfile'
              type: object
        "401":
          description: 未认证或令牌无效
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "404":
          description: 用户不存在
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "500":
          description: 服务器内部错误
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
      security:
      - BearerAuth: []
      summary: 获取用户资料
      tags:
      - Authentication
  /api/v1/auth/refresh-token:
    post:
      consumes:
      - application/json
      description: 使用刷新令牌获取新的访问令牌。当访问令牌过期时，前端可以使用此接口通过刷新令牌重新获取新的访问令牌和刷新令牌，无需重新进行钱包签名认证。
      parameters:
      - description: 刷新令牌请求体
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 刷新成功，返回新的访问令牌和刷新令牌
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/types.WalletConnectResponse'
              type: object
        "400":
          description: 请求参数错误
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "401":
          description: 刷新令牌无效或已过期
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "404":
          description: 用户不存在
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "500":
          description: 服务器内部错误
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
      summary: 刷新访问令牌
      tags:
      - Authentication
  /api/v1/auth/wallet-connect:
    post:
      consumes:
      - application/json
      description: 通过钱包签名进行用户认证。前端需要先让用户用钱包对特定消息进行签名，然后将钱包地址、签名和消息发送到此接口进行验证。验证成功后返回JWT访问令牌和刷新令牌。钱包地址必须是有效以太坊地址（长度42，0x前缀）。
      parameters:
      - description: 钱包连接认证请求体
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/types.WalletConnectRequest'
      produces:
      - application/json
      responses:
        "200":
          description: 认证成功，返回访问令牌和用户信息
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/types.WalletConnectResponse'
              type: object
        "400":
          description: 请求参数错误（INVALID_WALLET_ADDRESS等）
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "401":
          description: 认证失败，可能是签名无效或地址恢复失败
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
        "500":
          description: 服务器内部错误
          schema:
            allOf:
            - $ref: '#/definitions/types.APIResponse'
            - properties:
                error:
                  $ref: '#/definitions/types.APIError'
              type: object
      summary: 钱包连接认证
      tags:
      - Authentication
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and JWT token.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"